Move from bad half-ECS to better inheritance model
Since PhysicsComponents weren’t being stored in contiguous memory
anyway, they may as well become directly owned by PhysicsEntitys.
Other changes are related, and try to use fewer pointers where possible.
Both MainLevel and PhysicsManager now hold a reference to the Player,
which in the context of MainLevel is guaranteed to be
PhysicsManager::physEntities[0]. In the context of GameOver, the Player is
erased from that position, so the reference to the Player in
GameOver::m_physMan should not be used.